+++ mkosi/config.py
@@ -1215,6 +1215,7 @@ class ConfigSetting:
 
     # backward compatibility
     compat_names: tuple[str, ...] = ()
+    compat_longs: tuple[str, ...] = ()
 
     def __post_init__(self) -> None:
         if not self.name:
@@ -1425,7 +1426,7 @@ class Config:
     repository_key_fetch: bool
     repositories: list[str]
     cacheonly: Cacheonly
-    package_manager_trees: list[ConfigTree]
+    sandbox_trees: list[ConfigTree]
 
     output_format: OutputFormat
     manifest_format: list[ManifestFormat]
@@ -1543,7 +1544,7 @@ class Config:
     tools_tree_release: Optional[str]
     tools_tree_mirror: Optional[str]
     tools_tree_repositories: list[str]
-    tools_tree_package_manager_trees: list[ConfigTree]
+    tools_tree_sandbox_trees: list[ConfigTree]
     tools_tree_packages: list[str]
     tools_tree_certificates: bool
     runtime_trees: list[ConfigTree]
@@ -2010,15 +2011,15 @@ SETTINGS = (
         scope=SettingScope.universal,
     ),
     ConfigSetting(
-        dest="package_manager_trees",
-        long="--package-manager-tree",
+        dest="sandbox_trees",
+        long="--sandbox-tree",
+        compat_names=("PackageManagerTrees",),
+        compat_longs=("--package-manager-tree",),
         metavar="PATH",
         section="Distribution",
         parse=config_make_list_parser(delimiter=",", parse=make_tree_parser()),
-        default_factory=lambda ns: ns.skeleton_trees,
-        default_factory_depends=("skeleton_trees",),
-        help="Use a package manager tree to configure the package manager",
-        paths=("mkosi.pkgmngr", "mkosi.pkgmngr.tar",),
+        help="Use a sandbox tree to configure the various tools that mkosi executes",
+        paths=("mkosi.sandbox", "mkosi.sandbox.tar", "mkosi.pkgmngr", "mkosi.pkgmngr.tar",),
         scope=SettingScope.universal,
     ),
 
@@ -2932,12 +2933,14 @@ SETTINGS = (
         help="Repositories to use for the default tools tree",
     ),
     ConfigSetting(
-        dest="tools_tree_package_manager_trees",
-        long="--tools-tree-package-manager-tree",
+        dest="tools_tree_sandbox_trees",
+        long="--tools-tree-sandbox-tree",
+        compat_names=("ToolsTreePackageManagerTrees",),
+        compat_longs=("--tools-tree-package-manager-tree",),
         metavar="PATH",
         section="Host",
         parse=config_make_list_parser(delimiter=",", parse=make_tree_parser()),
-        help="Package manager trees for the default tools tree",
+        help="Sandbox trees for the default tools tree",
     ),
     ConfigSetting(
         dest="tools_tree_packages",
@@ -3348,18 +3351,19 @@ def create_argument_parser(chdir: bool = True) -> argparse.ArgumentParser:
             group = parser.add_argument_group(f"{s.section} configuration options")
             last_section = s.section
 
-        opts = [s.short, s.long] if s.short else [s.long]
+        for long in [s.long, *s.compat_longs]:
+            opts = [s.short, long] if s.short and long == s.long else [long]
 
-        group.add_argument(    # type: ignore
-            *opts,
-            dest=s.dest,
-            choices=s.choices,
-            metavar=s.metavar,
-            nargs=s.nargs,     # type: ignore
-            const=s.const,
-            help=s.help,
-            action=ConfigAction,
-        )
+            group.add_argument(    # type: ignore
+                *opts,
+                dest=s.dest,
+                choices=s.choices,
+                metavar=s.metavar,
+                nargs=s.nargs,     # type: ignore
+                const=s.const,
+                help=s.help if long == s.long else argparse.SUPPRESS,
+                action=ConfigAction,
+            )
 
     return parser
 
@@ -4127,7 +4131,7 @@ def summary(config: Config) -> str:
               Fetch Repository Keys: {yes_no(config.repository_key_fetch)}
                        Repositories: {line_join_list(config.repositories)}
              Use Only Package Cache: {config.cacheonly}
-              Package Manager Trees: {line_join_list(config.package_manager_trees)}
+                      Sandbox Trees: {line_join_list(config.sandbox_trees)}
 
     {bold("OUTPUT")}:
                       Output Format: {config.output_format}
@@ -4256,7 +4260,7 @@ def summary(config: Config) -> str:
                  Tools Tree Release: {none_to_none(config.tools_tree_release)}
                   Tools Tree Mirror: {none_to_default(config.tools_tree_mirror)}
             Tools Tree Repositories: {line_join_list(config.tools_tree_repositories)}
-   Tools Tree Package Manager Trees: {line_join_list(config.tools_tree_package_manager_trees)}
+           Tools Tree Sandbox Trees: {line_join_list(config.tools_tree_sandbox_trees)}
                 Tools Tree Packages: {line_join_list(config.tools_tree_packages)}
             Tools Tree Certificates: {yes_no(config.tools_tree_certificates)}
                       Runtime Trees: {line_join_list(config.runtime_trees)}